+++ src/commands/build.rs
@@ -107,7 +107,7 @@ impl Command<Context> for Build {
             // read config.toml for setting any env variables
             .from_config(c.get_config())?
             // read ip manifest for env variables
-            .from_ip(&Ip::load(c.get_ip_path().unwrap().clone())?)?
+            .from_ip(&Ip::load(c.get_ip_path().unwrap().clone(), true)?)?
             .add(EnvVar::new().key(ORBIT_BLUEPRINT).value(BLUEPRINT_FILE))
             .add(EnvVar::new().key(ORBIT_BUILD_DIR).value(b_dir))
             .initialize();