+++ src/commands/env.rs
@@ -86,7 +86,7 @@ impl Command<Context> for Env {
         // check if in an ip to add those variables
         if let Some(ip_path) = c.get_ip_path() {
             // check ip
-            if let Ok(ip) = Ip::load(ip_path.clone()) {
+            if let Ok(ip) = Ip::load(ip_path.clone(), true) {
                 env = env.from_ip(&ip)?;
             }
             // check the build directory