+++ src/util/filesystem.rs
@@ -1,7 +1,6 @@
 use crate::core::fileset;
 use crate::core::lockfile;
 use crate::core::manifest;
-use crate::core::pubfile::PubFile;
 use fs_extra;
 use home::home_dir;
 use ignore::WalkBuilder;
@@ -23,7 +22,7 @@ use super::anyerror::Fault;
 /// final [String] entries in the resulting vector.
 ///
 /// Ignores ORBIT_SUM_FILE, .git directory, ORBIT_METADATA_FILE, and IP_LOCK_FILE.
-pub fn gather_current_files(path: &PathBuf, strip_base: bool, use_pub_file: bool) -> Vec<String> {
+pub fn gather_current_files(path: &PathBuf, strip_base: bool) -> Vec<String> {
     let m = WalkBuilder::new(path)
         .hidden(false)
         .add_custom_ignore_filename(ORBIT_IGNORE_FILE)
@@ -52,27 +51,6 @@ pub fn gather_current_files(path: &PathBuf, strip_base: bool, use_pub_file: bool
             }
         })
         .collect();
-    // determine what HDL files should be kept (.orbitpub acts as opposite of ignore)
-    if use_pub_file == true {
-        let pub_filepath = path.join(PubFile::get_filename());
-        let pub_file = PubFile::new(&pub_filepath);
-        files = files
-            .into_iter()
-            .filter_map(|p| {
-                match fileset::is_hdl(p.as_ref()) {
-                    true => {
-                        // the pub file explicitly lists the files to keep
-                        match pub_file.is_included(&p) {
-                            true => Some(p),
-                            false => None,
-                        }
-                    }
-                    // keep all non-hdl files
-                    false => Some(p),
-                }
-            })
-            .collect();
-    }
     // sort the fileset for reproducibility purposes
     files.sort();
     files